+++ test/ck_tile/gemm/test_gemm_pipeline_util.hpp
@@ -138,9 +138,12 @@ class TestCkTileGemmPipeline : public ::testing::Test
         const bool has_hot_loop            = BaseGemmPipeline::BlockHasHotloop(num_loop);
         const ck_tile::TailNumber tail_num = BaseGemmPipeline::GetBlockLoopTailNum(num_loop);
 
-        const auto Run = [&](const auto has_hot_loop_, const auto tail_number_) {
-            constexpr bool has_hot_loop_v = has_hot_loop_.value;
-            constexpr auto tail_number_v  = tail_number_.value;
+        const auto Run = [&](const auto has_hot_loop_,
+                             const auto tail_number_,
+                             const auto memory_operation_) {
+            constexpr bool has_hot_loop_v   = has_hot_loop_.value;
+            constexpr auto tail_number_v    = tail_number_.value;
+            constexpr auto memory_operation = memory_operation_.value;
 
             using UniversalGemmProblem = ck_tile::UniversalGemmPipelineProblem<ADataType,
                                                                                BDataType,
@@ -168,7 +171,8 @@ class TestCkTileGemmPipeline : public ::testing::Test
                                                  M_Warp_Tile,
                                                  N_Warp_Tile,
                                                  K_Warp_Tile,
-                                                 UniversalGemmProblem::TransposeC>>;
+                                                 UniversalGemmProblem::TransposeC,
+                                                 memory_operation>>;
 
             using Kernel = ck_tile::GemmKernel<TilePartitioner, GemmPipeline, GemmEpilogue>;
             auto kargs   = Kernel::MakeKernelArgs(args);
@@ -193,15 +197,32 @@ class TestCkTileGemmPipeline : public ::testing::Test
                 s, ck_tile::make_kernel<blocks.x, kBlockPerCu>(Kernel{}, grids, blocks, 0, kargs));
         };
 
+        const auto RunSplitk = [&](const auto has_hot_loop_, const auto tail_number_) {
+            if(args.k_batch == 1)
+            {
+                Run(has_hot_loop_,
+                    tail_number_,
+                    ck_tile::integral_constant<ck_tile::memory_operation_enum,
+                                               ck_tile::memory_operation_enum::set>{});
+            }
+            else
+            {
+                Run(has_hot_loop_,
+                    tail_number_,
+                    ck_tile::integral_constant<ck_tile::memory_operation_enum,
+                                               ck_tile::memory_operation_enum::atomic_add>{});
+            }
+        };
+
         if(has_hot_loop)
         {
             if constexpr(PipelineType == GemmPipelineType::CompV3)
             {
                 if(tail_num == ck_tile::TailNumber::Full)
                 {
-                    Run(ck_tile::bool_constant<true>{},
-                        ck_tile::integral_constant<ck_tile::TailNumber,
-                                                   ck_tile::TailNumber::Full>{});
+                    RunSplitk(ck_tile::bool_constant<true>{},
+                              ck_tile::integral_constant<ck_tile::TailNumber,
+                                                         ck_tile::TailNumber::Full>{});
                 }
                 else
                 {
@@ -219,69 +240,69 @@ class TestCkTileGemmPipeline : public ::testing::Test
                 // Tail pipeline One to Seven
                 if(tail_num == ck_tile::TailNumber::One)
                 {
-                    Run(ck_tile::bool_constant<true>{},
-                        ck_tile::integral_constant<ck_tile::TailNumber,
-                                                   ck_tile::TailNumber::One>{});
+                    RunSplitk(ck_tile::bool_constant<true>{},
+                              ck_tile::integral_constant<ck_tile::TailNumber,
+                                                         ck_tile::TailNumber::One>{});
                 }
                 else if(tail_num == ck_tile::TailNumber::Full)
                 {
-                    Run(ck_tile::bool_constant<true>{},
-                        ck_tile::integral_constant<ck_tile::TailNumber,
-                                                   ck_tile::TailNumber::Full>{});
+                    RunSplitk(ck_tile::bool_constant<true>{},
+                              ck_tile::integral_constant<ck_tile::TailNumber,
+                                                         ck_tile::TailNumber::Full>{});
                 }
 
                 if constexpr(BaseGemmPipeline::PrefetchStages > 2)
                 {
                     if(tail_num == ck_tile::TailNumber::Two)
                     {
-                        Run(ck_tile::bool_constant<true>{},
-                            ck_tile::integral_constant<ck_tile::TailNumber,
-                                                       ck_tile::TailNumber::Two>{});
+                        RunSplitk(ck_tile::bool_constant<true>{},
+                                  ck_tile::integral_constant<ck_tile::TailNumber,
+                                                             ck_tile::TailNumber::Two>{});
                     }
                 }
                 if constexpr(BaseGemmPipeline::PrefetchStages > 3)
                 {
                     if(tail_num == ck_tile::TailNumber::Three)
                     {
-                        Run(ck_tile::bool_constant<true>{},
-                            ck_tile::integral_constant<ck_tile::TailNumber,
-                                                       ck_tile::TailNumber::Three>{});
+                        RunSplitk(ck_tile::bool_constant<true>{},
+                                  ck_tile::integral_constant<ck_tile::TailNumber,
+                                                             ck_tile::TailNumber::Three>{});
                     }
                 }
                 if constexpr(BaseGemmPipeline::PrefetchStages > 4)
                 {
                     if(tail_num == ck_tile::TailNumber::Four)
                     {
-                        Run(ck_tile::bool_constant<true>{},
-                            ck_tile::integral_constant<ck_tile::TailNumber,
-                                                       ck_tile::TailNumber::Four>{});
+                        RunSplitk(ck_tile::bool_constant<true>{},
+                                  ck_tile::integral_constant<ck_tile::TailNumber,
+                                                             ck_tile::TailNumber::Four>{});
                     }
                 }
                 if constexpr(BaseGemmPipeline::PrefetchStages > 5)
                 {
                     if(tail_num == ck_tile::TailNumber::Five)
                     {
-                        Run(ck_tile::bool_constant<true>{},
-                            ck_tile::integral_constant<ck_tile::TailNumber,
-                                                       ck_tile::TailNumber::Five>{});
+                        RunSplitk(ck_tile::bool_constant<true>{},
+                                  ck_tile::integral_constant<ck_tile::TailNumber,
+                                                             ck_tile::TailNumber::Five>{});
                     }
                 }
                 if constexpr(BaseGemmPipeline::PrefetchStages > 6)
                 {
                     if(tail_num == ck_tile::TailNumber::Six)
                     {
-                        Run(ck_tile::bool_constant<true>{},
-                            ck_tile::integral_constant<ck_tile::TailNumber,
-                                                       ck_tile::TailNumber::Six>{});
+                        RunSplitk(ck_tile::bool_constant<true>{},
+                                  ck_tile::integral_constant<ck_tile::TailNumber,
+                                                             ck_tile::TailNumber::Six>{});
                     }
                 }
                 if constexpr(BaseGemmPipeline::PrefetchStages > 7)
                 {
                     if(tail_num == ck_tile::TailNumber::Seven)
                     {
-                        Run(ck_tile::bool_constant<true>{},
-                            ck_tile::integral_constant<ck_tile::TailNumber,
-                                                       ck_tile::TailNumber::Seven>{});
+                        RunSplitk(ck_tile::bool_constant<true>{},
+                                  ck_tile::integral_constant<ck_tile::TailNumber,
+                                                             ck_tile::TailNumber::Seven>{});
                     }
                 }
             }
@@ -290,15 +311,15 @@ class TestCkTileGemmPipeline : public ::testing::Test
             {
                 if(tail_num == ck_tile::TailNumber::Three)
                 {
-                    Run(ck_tile::bool_constant<true>{},
-                        ck_tile::integral_constant<ck_tile::TailNumber,
-                                                   ck_tile::TailNumber::Three>{});
+                    RunSplitk(ck_tile::bool_constant<true>{},
+                              ck_tile::integral_constant<ck_tile::TailNumber,
+                                                         ck_tile::TailNumber::Three>{});
                 }
                 else
                 {
-                    Run(ck_tile::bool_constant<true>{},
-                        ck_tile::integral_constant<ck_tile::TailNumber,
-                                                   ck_tile::TailNumber::Two>{});
+                    RunSplitk(ck_tile::bool_constant<true>{},
+                              ck_tile::integral_constant<ck_tile::TailNumber,
+                                                         ck_tile::TailNumber::Two>{});
                 }
             }
         }
@@ -307,7 +328,8 @@ class TestCkTileGemmPipeline : public ::testing::Test
             // Tail number always Full - #PrefetchStages
             if(tail_num == ck_tile::TailNumber::Full)
             {
-                Run(ck_tile::bool_constant<false>{},
+                RunSplitk(
+                    ck_tile::bool_constant<false>{},
                     ck_tile::integral_constant<ck_tile::TailNumber, ck_tile::TailNumber::Full>{});
             }
             else